Fix operation on 32 bit platforms
The issue might be repdroduced by compiling OpenFEC with -m32 flag,
or compiling and calling it on ARM. Then OpenFEC's tests fails.

The issue is in the file of_symbol.c in symbols arithmetic operations.
According to OpenFEC's tests (make test) the issue is in
of_add_to_multiple_symbols(...) and of_add_from_multiple_symbols(...).